requirements.txt
Flask==2.0.2
werkzeug==2.0.2
langchain
spacy>=3.0.0,<3.1.0
langchain-google-genai
google-generativeai
waitress==2.1.0
python-dotenv==0.19.2
flask-cors
faiss-cpu
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz